+++ CountlyConfig.h
@@ -676,9 +676,9 @@ typedef enum : NSUInteger
 @property(nonatomic, copy) NSString *sdkBehaviorSettings;
 
 /**
- * Disable the server configuration
+ * Disable the server configuration updates to the server
  */
-@property(nonatomic) BOOL disableSDKBehaviorSettings;
+@property(nonatomic) BOOL disableSDKBehaviorSettingsUpdates;
 
 /**
  * Will disable back off mechanism