+++ CountlyServerConfig.m
@@ -44,7 +44,7 @@
 @property (nonatomic) NSInteger version;
 @property (nonatomic) long long timestamp;
 @property (nonatomic) long long lastFetchTimestamp;
-@property (nonatomic) BOOL serverConfigDisabled;
+@property (nonatomic) BOOL serverConfigUpdatesDisabled;
 
 @end
 
@@ -108,7 +108,7 @@ NSString *const kRBOMDuration = @"bom_d";
         _version = 0;
         _currentServerConfigUpdateInterval = 4;
         _requestTimer = nil;
-        _serverConfigDisabled = NO;
+        _serverConfigUpdatesDisabled = NO;
         [self setDefaultValues];
     }
     return self;
@@ -116,9 +116,6 @@ NSString *const kRBOMDuration = @"bom_d";
 
 - (void)retrieveServerConfigFromStorage:(NSString *)sdkBehaviorSettings
 {
-    if (_serverConfigDisabled) {
-        return;
-    }
     NSError *error = nil;
     NSDictionary *serverConfigObject = [CountlyPersistency.sharedInstance retrieveServerConfig];
     if (serverConfigObject.count == 0 && sdkBehaviorSettings)
@@ -217,10 +214,6 @@ NSString *const kRBOMDuration = @"bom_d";
 
 - (void)notifySdkConfigChange:(CountlyConfig *)config
 {
-    if (_serverConfigDisabled) {
-        return;
-    }
-    
     config.enableDebug = _loggingEnabled || config.enableDebug;
     CountlyCommon.sharedInstance.enableDebug = config.enableDebug;
 
@@ -325,7 +318,7 @@ NSString *const kRBOMDuration = @"bom_d";
 
 - (void)fetchServerConfigIfTimeIsUp
 {
-    if (_serverConfigDisabled) {
+    if (_serverConfigUpdatesDisabled) {
         return;
     }
     
@@ -343,10 +336,12 @@ NSString *const kRBOMDuration = @"bom_d";
 
 - (void)fetchServerConfig:(CountlyConfig *)config
 {
-    if (_serverConfigDisabled) {
+    CLY_LOG_D(@"%s, fetching sdk behavior settings", __FUNCTION__);
+    
+    if (_serverConfigUpdatesDisabled) {
+        CLY_LOG_D(@"%s, sdk behavior settings updates disabled, omitting fetch", __FUNCTION__);
         return;
     }
-    CLY_LOG_D(@"Fetching server configs...");
     
     if (CountlyDeviceInfo.sharedInstance.isDeviceIDTemporary)
         return;
@@ -445,7 +440,7 @@ NSString *const kRBOMDuration = @"bom_d";
 }
 
 - (void)disableSDKBehaviourSettings {
-    _serverConfigDisabled = YES;
+    _serverConfigUpdatesDisabled = YES;
 }
 
 - (BOOL)trackingEnabled